requirements.txt
kafka-python
pandas
clickhouse-driver
pyspark==3.5.0
